refactor: move retry queue module to common
