+++ packages/analytics-js/.size-limit.mjs
@@ -30,24 +30,24 @@ export default [
     name: 'Core - Modern - NPM (ESM)',
     path: 'dist/npm/modern/esm/index.mjs',
     import: '*',
-    limit: '27 KiB',
+    limit: '30 KiB',
   },
   {
     name: 'Core - Modern - NPM (CJS)',
     path: 'dist/npm/modern/cjs/index.cjs',
     import: '*',
-    limit: '27.1 KiB',
+    limit: '30 KiB',
   },
   {
     name: 'Core - Modern - NPM (UMD)',
     path: 'dist/npm/modern/umd/index.js',
     import: '*',
-    limit: '27 KiB',
+    limit: '30 KiB',
   },
   {
     name: 'Core - Modern - CDN',
     path: 'dist/cdn/modern/iife/rsa.min.js',
-    limit: '27.5 KiB',
+    limit: '30 KiB',
   },
   {
     name: 'Core (Bundled) - Legacy - NPM (ESM)',
